third_party/blink/web_tests/external/wpt/navigation-api/ordering-and-transition/navigate-same-document-intercept-reject_no-currententrychange-expected.txt

This is a testharness.js-based test.
[FAIL] event and promise ordering for same-document navigation.navigate() intercepted by passing a rejected promise to intercept()
  assert_array_equals: expected property 2 to be "AbortSignal abort" but got "committed fulfilled" (expected array ["navigate", "handler run", "AbortSignal abort", "navigateerror", "committed fulfilled", "transition.committed fulfilled", "promise microtask", "finished rejected", "transition.finished rejected"] got ["navigate", "handler run", "committed fulfilled", "transition.committed fulfilled", "promise microtask", "AbortSignal abort", "navigateerror", "finished rejected", "transition.finished rejected"])
Harness: the test ran to completion.
